backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
opencv-python==4.8.1.78
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
pillow==10.0.1
python-dotenv==1.0.0
mlflow==2.8.1
cryptography==41.0.4
pytest==7.4.2
flake8==6.1.0
black==23.9.1
